pyproject.toml
[tool.poetry]
name = "prop-mgmt-backend"
version = "0.1.0"
description = "Property Management App — Backend API"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
google-cloud-bigquery = "^3.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
